stm32l4-multi: Moved libtty uart implementation to one file
